[Refactor] (code) Move the crawler module to sub-package 'crawler' and rename it as 'cluster'.
