+++ docs/smoothcrawler_cluster/api_references/crawler/abstract_crawler.rst
@@ -5,12 +5,19 @@ Abstract Crawlers
 Base Distributed Crawler
 ==========================
 
-.. autoclass:: smoothcrawler_cluster.crawler.BaseDistributedCrawler
+.. autoclass:: smoothcrawler_cluster.crawler.cluster.BaseDistributedCrawler
+    :members:
+
+
+Base Cluster Crawler
+=====================
+
+.. autoclass:: smoothcrawler_cluster.crawler.cluster.BaseClusterCrawler
     :members:
 
 
 Base Decentralized Crawler
 ============================
 
-.. autoclass:: smoothcrawler_cluster.crawler.BaseDecentralizedCrawler
+.. autoclass:: smoothcrawler_cluster.crawler.cluster.BaseDecentralizedCrawler
     :members:
