+++ docs/smoothcrawler_cluster/api_references/crawler/zookeeper_crawler.rst
@@ -2,6 +2,6 @@
 Zookeeper Crawlers
 ====================
 
-.. autoclass:: smoothcrawler_cluster.crawler.ZookeeperCrawler
+.. autoclass:: smoothcrawler_cluster.crawler.cluster.ZookeeperCrawler
     :private-members: _is_ready_by_groupstate, _update_crawler_role, _run_updating_heartbeat_thread, _update_heartbeat, _run_crawling_processing, _chk_register
     :members:
